backend/src/pybr25/core/setuphandlers/__init__.py
from Products.CMFPlone.interfaces import INonInstallable
from zope.interface import implementer


@implementer(INonInstallable)
class HiddenProfiles:
    def getNonInstallableProfiles(self):
        """Hide initial profile from site-creation and quickinstaller."""
        return [
            "pybr25.core:initial",
        ]

    def getNonInstallableProducts(self):
        """Hide the upgrades package from site-creation and quickinstaller."""
        return [
            "pybr25.core.upgrades",
        ]
